pyproject.toml
[tool.poetry]
name = "repl_nix_PremBot"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^2.1.0"
requests = "^2.28.1"
discord-pretty-help = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
